Adds new ResolvesAuthorizable contract.
Allows non Authorizable class instances to be passed into the policy()
helper function. The contract requires a resolveAuthorizable() method to
exist, serving up an Authorizable class.

This allows, for example, a presenter class to be passed directly to the
policy() helper in a view. The presenter can be configured to resolve
the related model instance that can be authorized against.

m